+++ geocoding_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.0
+
+- **Breaking Change** Matches the changes to the platform interface calls, the locale is now set in a separate call.
+- Exposes the full geocoder API, making it possible to use Android-specific logic.
+
 ## 2.1.0
 
 * Splits from `geocoding` as a federated implementation.